Moves AudioPlayer to its own folder
