+++ src/app/app.module.ts
@@ -186,7 +186,7 @@ import { NavigationService } from './services/navigation/navigation.service';
 import { NowPlayingNavigationService } from './services/now-playing-navigation/now-playing-navigation.service';
 import { PlaybackIndicationService } from './services/playback-indication/playback-indication.service';
 import { PlaybackInformationService } from './services/playback-information/playback-information.service';
-import { AudioPlayer } from './services/playback/audio-player';
+import { AudioPlayer } from './services/playback/audio-player/audio-player';
 import { ProgressUpdater } from './services/playback/progress-updater';
 import { PlaylistFolderModelFactory } from './services/playlist-folder/playlist-folder-model-factory';
 import { PlaylistFolderService } from './services/playlist-folder/playlist-folder.service';
@@ -211,7 +211,7 @@ import { OnlineLyricsGetter } from './services/lyrics/online-lyrics-getter';
 import { IntegrationTestRunner } from './testing/integration-test-runner';
 import { EventListenerService } from './services/event-listener/event-listener.service';
 import { Desktop } from './common/io/desktop';
-import { AudioPlayerBase } from './services/playback/audio-player.base';
+import { AudioPlayerBase } from './services/playback/audio-player/audio-player.base';
 import { EventListenerServiceBase } from './services/event-listener/event-listener.service.base';
 import { LyricsServiceBase } from './services/lyrics/lyrics.service.base';
 import { ArtistInformationServiceBase } from './services/artist-information/artist-information.service.base';
